Polish election data 2015

Polish Presidential election 2015 data by gminy and Warsaw borough areal units

pol_pres15

Format

sf data frame object with 2495 areal units and 65 variables

TERYT, TERYT0, gm0TERYT areal unit IDs

name0original areal unit names

namecleaned areal unit names

typesfactor with levels “Rural”, “Urban”, “Urban/rural” and “Warsaw borough”

I_turnoutFirst round turnout proportion

II_turnoutRunoff round turnout proportion

I_Duda_shareWinner first round share

II_Duda_shareWinner runoff round share

I_Komorowski_shareIncumbent first round share

II_Komorowski_shareIncumbent runoff round share

I_*First round aggregated counts of all polling station data

II_*Runoff round aggregated counts of all polling station data

Source

https://prezydent2015.pkw.gov.pl/319_Pierwsze_glosowanie, https://prezydent2015.pkw.gov.pl/325_Ponowne_glosowanie, http://www.codgik.gov.pl/index.php/darmowe-dane/prg.html

Note

“PVE” in variable names means “postal voting envelopes”; voters requesting a postal voting package are expected to return a postal voting envelope with a declaration, and a sealed voting envelope to be placed in the ballot box.

Author

Roger Bivand

Examples

if (FALSE) {
data("pol_pres15", package = "spDataLarge")
wd = aggregate(pol_pres15$I_entitled_to_vote, list(pol_pres15$types), sum)$x
boxplot(I_turnout ~ types, data = pol_pres15, width = wd)
}